separates out game logic from index file
